Extract inline page components from App.tsx into playground/src/pages/
Agent-Logs-Url: https://github.com/SergeiGolos/wod-wiki/sessions/5834bf9f-04ef-44da-a67d-c8bdbf0c4daa
