+++ playground/src/pages/CollectionsPage.tsx
@@ -1,0 +1,4 @@
+/**
+ * CollectionsPage — re-exported from views/CollectionsPage for the pages/ module.
+ */
+export { CollectionsPage } from '../views/CollectionsPage'
