+++ playground/src/pages/HomeView.tsx
@@ -1,0 +1,5 @@
+/**
+ * HomeView page — re-exported from views/HomeView for the pages/ module.
+ */
+export { HomeView } from '../views/HomeView'
+export type { HomeViewProps } from '../views/HomeView'
