+++ playground/src/pages/JournalWeeklyPage.tsx
@@ -1,0 +1,4 @@
+/**
+ * JournalWeeklyPage — re-exported from views/ListViews for the pages/ module.
+ */
+export { JournalWeeklyPage } from '../views/ListViews'
